+++ script/ordered_priority_queue_hw.tcl
@@ -8,7 +8,7 @@ package require -exact altera_terp 1.0
 
 set_module_property NAME                             ordered_priority_queue
 set_module_property DISPLAY_NAME                     "Ordered Priority Queue"
-set_module_property VERSION                          26.4.12.0428
+set_module_property VERSION                          26.4.13.0428
 set_module_property DESCRIPTION                      "Ordered Priority Queue Mu3e IP Core"
 set_module_property GROUP                            "Mu3e Data Plane/Modules"
 set_module_property AUTHOR                           "Yifeng Wang (yifenwan@phys.ethz.ch)"
@@ -311,7 +311,7 @@ proc opq_preset_summary_html {selected_preset} {
 set IP_UID_DEFAULT_CONST        1330663757
 set VERSION_MAJOR_DEFAULT_CONST 26
 set VERSION_MINOR_DEFAULT_CONST 4
-set VERSION_PATCH_DEFAULT_CONST 12
+set VERSION_PATCH_DEFAULT_CONST 13
 set BUILD_DEFAULT_CONST         428
 set VERSION_DATE_DEFAULT_CONST  20260428
 # 32-bit packaged provenance stamp for this release family